Removed dynamic spring arm length change logic and connected variables in GliderPawn, due to the jittering effect and the fact that camera lag achieves the same results.
Added logic, which allows to initiate camera lag enabling or disabling. This was needed to handle issue, when glider is pushed by booster objects via high impulse and camera can't keep up. This is also handled by interpolation in order to achieve smooth transition from disabled to enabled state. This state is being enabled and disabled by booster object itself.

Honestly, this is very odd way to solve this issue and I find it a little bit hardcoded. In case, we find something better it should be applied instead. Also, it seems like jittering effect is still intact, but not that strong as it was before.